# backend_go/docs/swagger.yaml
basePath: /api
definitions:
  gateways.Invoice:
    properties:
      gateway_invoice_id:
        description: The ID of the invoice in the gateway's system
        type: string
      pay_url:
        description: The URL the user should be redirected to for payment
        type: string
    type: object
...
  handlers.createInvoicePayload:
    properties:
      amount:
        type: number
      bot_user_id:
        type: integer
      gateway_name:
        type: string
    required:
    - amount
    - bot_user_id
    - gateway_name
    type: object
...
  handlers.gatewayDTO:
    properties:
      display_name:
        type: string
      name:
        type: string
    type: object
...
  handlers.productCreatePayload:
    properties:
      category_id:
        type: integer
      initial_stock:
        minimum: 0
        type: integer
      name:
        type: string
      price:
        minimum: 0
        type: number
      subscription_period_days:
        minimum: 0
        type: integer
      type:
        enum:
        - item
        - subscription
        type: string
    required:
    - category_id
    - name
    type: object
  handlers.productUpdatePayload:
    properties:
      category_id:
        type: integer
      name:
        type: string
      price:
        minimum: 0
        type: number
      subscription_period_days:
        minimum: 0
        type: integer
      type:
        enum:
        - item
        - subscription
        type: string
    required:
    - category_id
    - name
    type: object
...
  handlers.stockMovementPayload:
    properties:
      description:
        type: string
      order_id:
        type: integer
      quantity:
        type: integer
      type:
        $ref: '#/definitions/models.StockMovementType'
    required:
    - quantity
    - type
    type: object
...
  models.ProductResponse:
    properties:
      category_id:
        type: integer
      external_id:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      provider:
        type: string
      stock:
        type: integer
      subscription_period_days:
        type: integer
      type:
        type: string
    type: object
  models.StockMovement:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      orderID:
        type: integer
      productID:
        type: integer
      quantity:
        type: integer
      type:
        $ref: '#/definitions/models.StockMovementType'
    type: object
  models.StockMovementType:
    enum:
    - initial
    - sale
    - restock
    - return
    type: string
    x-enum-varnames:
    - Initial
    - Sale
    - Restock
    - Return
...
  responses.ErrorResponseSchema:
    description: This is the standard error response structure.
    properties:
      data: {}
      error:
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  responses.ResponseSchema-array_handlers_gatewayDTO:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.gatewayDTO'
        type: array
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.ResponseSchema-array_models_ProductResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.ProductResponse'
        type: array
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.ResponseSchema-gateways_Invoice:
    properties:
      data:
        $ref: '#/definitions/gateways.Invoice'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.ResponseSchema-models_ProductResponse:
    properties:
      data:
        $ref: '#/definitions/models.ProductResponse'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
  responses.ResponseSchema-models_StockMovement:
    properties:
      data:
        $ref: '#/definitions/models.StockMovement'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for your project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Your Project API
  version: "1.0"
paths:
...
  /deposit/invoice:
    post:
      consumes:
      - application/json
      description: Creates a new payment invoice for a selected gateway and amount.
      parameters:
      - description: Invoice creation data
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/handlers.createInvoicePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.ResponseSchema-gateways_Invoice'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: Create Payment Invoice
      tags:
      - Payments
  /gateways:
    get:
      description: Retrieves a list of available payment gateways.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-array_handlers_gatewayDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: Get Payment Gateways
      tags:
      - Payments
...
  /products:
    get:
      description: Get a list of all products, optionally filtered by category.
      parameters:
      - collectionFormat: csv
        description: Category IDs to filter by
        in: query
        items:
          type: string
        name: category_ids[]
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-array_models_ProductResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      - ServiceApiKeyAuth: []
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Adds a new product to the store.
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handlers.productCreatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.ResponseSchema-models_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Create a product
      tags:
      - Products
  /products/{id}:
    delete:
      description: Deletes a product by its ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      description: Retrieves details for a single product.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-models_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Get a product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Updates an existing product.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handlers.productUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-models_ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Update a product
      tags:
      - Products
  /products/{id}/stock/movements:
    post:
      consumes:
      - application/json
      description: Adds a stock movement (e.g., initial stock, sale, return) for a
        product.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Stock movement data
        in: body
        name: movement
        required: true
        schema:
          $ref: '#/definitions/handlers.stockMovementPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.ResponseSchema-models_StockMovement'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Create a stock movement
      tags:
      - Products
      - Stock
...
  /webhooks/{gateway_name}:
    post:
      consumes:
      - application/json
      description: Handles incoming webhook notifications from a payment gateway.
      parameters:
      - description: Gateway Name
        in: path
        name: gateway_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      summary: Handle Gateway Webhook
      tags:
      - Payments
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ServiceApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"